Remove query_jobs_buildername function.
